requirements.txt
python-dotenv
openai
streamlit
requests